Support decimal tire diameters for heavy-truck tires (17.5, 22.5)
tire_diameter was stored and validated as an integer, rejecting the
half-inch rim diameters used by heavy-truck tires (e.g. 295/80R22.5).
Migrates the column to decimal(4,1), relaxes validation to numeric,
and updates Stock::parseSearchQuery() to recognize decimal diameter
formats (295/80R22.5, 12R17.5, standalone 17.5) while preserving
integer output for standard whole-size diameters.

+++ back/app/Models/ProductTyre.php
@@ -30,7 +30,7 @@ class ProductTyre extends Model
     protected $casts = [
         'tire_width' => 'integer',
         'tire_height' => 'integer',
-        'tire_diameter' => 'integer',
+        'tire_diameter' => 'float',
         'tire_runflat' => 'boolean',
         'tire_reinforced' => 'boolean',
         'eu_noise_db' => 'integer',
